Manager: add `legacy` optional param to get_capacity
The `legacy` param will be passed directly to the underlying /capacity
call. This will instruct the api to return capacity status for legacy
server types (baremetal_*) and/or new *.*.$arch "class" based types. The
current default is `only` which returns only the `baremetal_*`. In the
not-so-distant-future the default will change to `exclude`, so only
*.*.$arch types will be returned.

This functionality was incorrectly merged in Feb, woops. Well at least
there was no release.